data-platform-spark-framework/pyproject.toml
[project]
name = "pipeline"
version = "1.0.2.dev1"
description = "PySpark jobs for Envilink ETL pipeline framework"
requires-python = ">=3.11"
dependencies = [
    "pyspark==3.5.3",
    "apache-sedona==1.8.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "google-cloud-bigquery>=3.11.0",
    "google-cloud-bigquery-storage>=2.19.0",
    "google-cloud-storage>=2.10.0",
    "great-expectations>=1.9.0",
    "dotenv>=0.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline"]

[tool.hatch.build.targets.wheel.force-include]
"src/pipeline/env_config.yaml" = "pipeline/env_config.yaml"

[tool.hatch.build]
include = [
    ".venv/lib/python3.11/site-packages/",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
